Reorganize guard validation code to follow implementation plan structure
